requirements.txt

Flask==3.0.0


mysql-connector-python==8.2.0


pytest==7.4.3
pytest-cov==4.1.0


locust==2.19.0


playwright==1.40.0
pytest-playwright==0.4.3